move inProgress modal to molecules, make more generic
