+++ app/src/molecules/InProgressModal/InProgressModal.tsx
@@ -1,32 +1,29 @@
 import * as React from 'react'
-import { useTranslation } from 'react-i18next'
-import { StyledText } from '../../atoms/text'
-import { WizardHeader } from '../../molecules/WizardHeader'
+import { WizardHeader } from '../WizardHeader'
 import {
   ALIGN_CENTER,
   COLORS,
   DIRECTION_COLUMN,
   Flex,
   Icon,
-  SPACING,
 } from '@opentrons/components'
 
 interface Props {
-  title: string
+  wizardHeaderTitle: string
   currentStep: number
   totalSteps: number
+  children?: JSX.Element
 }
 
 export function InProgressModal(props: Props): JSX.Element {
-  const { title, totalSteps, currentStep } = props
-  const { t } = useTranslation('change_pipette')
+  const { wizardHeaderTitle, totalSteps, currentStep, children } = props
 
   return (
     <>
       <WizardHeader
         totalSteps={totalSteps}
         currentStep={currentStep}
-        title={title}
+        title={wizardHeaderTitle}
       />
       <Flex
         alignItems={ALIGN_CENTER}
@@ -41,13 +38,7 @@ export function InProgressModal(props: Props): JSX.Element {
           aria-label="spinner"
           spin
         />
-        <StyledText
-          as="h1"
-          marginTop={SPACING.spacing5}
-          marginBottom={SPACING.spacing3}
-        >
-          {t('moving_gantry')}
-        </StyledText>
+        {children}
       </Flex>
     </>
   )
